remove appendix; change NPPS label; add brokenaxes
Better to not have Lambda(r) since we want discrete labels.

WRT the appendix:
As it was, the calculation that it's doing is roughly the same as what's
done in the main text.  However it tries to be more formal, and in
expanding the definition of "rate density" makes a mistake.

Defining Gamma(r,M) as number of planets per star per unit planet radius
per unit stellar mass is not actually the density function that we want.
This definition of Gamma is ill-defined because of the added measure in
stellar counts (the denominator). This problem is irrelevant to the main
text, in which we focus exclusively on planet radius. So it'll be a
question left for future work.